[INLONG-2545]agent pull task(binlog ,file ,kafka) from master
