+++ inlong-agent/agent-plugins/src/main/java/org/apache/inlong/agent/plugin/fetcher/ManagerFetcher.java
@@ -74,10 +74,10 @@ import org.apache.inlong.agent.conf.ProfileFetcher;
 import org.apache.inlong.agent.conf.TriggerProfile;
 import org.apache.inlong.agent.core.AgentManager;
 import org.apache.inlong.agent.db.CommandDb;
-import org.apache.inlong.agent.db.CommandEntity;
-import org.apache.inlong.agent.dto.CmdConfig;
-import org.apache.inlong.agent.dto.TaskRequestDto;
-import org.apache.inlong.agent.enums.ManagerOpEnum;
+import org.apache.inlong.commons.db.CommandEntity;
+import org.apache.inlong.commons.dto.CmdConfig;
+import org.apache.inlong.commons.dto.TaskRequestDto;
+import org.apache.inlong.commons.enums.ManagerOpEnum;
 import org.apache.inlong.agent.plugin.Trigger;
 import org.apache.inlong.agent.plugin.fetcher.dtos.ConfirmAgentIpRequest;
 import org.apache.inlong.agent.plugin.fetcher.dtos.DbCollectorTaskRequestDto;
