+++ inlong-agent/agent-plugins/src/main/java/org/apache/inlong/agent/plugin/fetcher/dtos/TaskResult.java
@@ -19,8 +19,8 @@ package org.apache.inlong.agent.plugin.fetcher.dtos;
 
 import lombok.Data;
 import org.apache.inlong.agent.conf.TriggerProfile;
-import org.apache.inlong.agent.dto.CmdConfig;
-import org.apache.inlong.agent.dto.DataConfig;
+import org.apache.inlong.commons.dto.CmdConfig;
+import org.apache.inlong.commons.dto.DataConfig;
 
 import java.util.ArrayList;
 import java.util.List;
