+++ inlong-common/src/main/java/org/apache/inlong/commons/dto/TaskRequestDto.java
@@ -15,23 +15,16 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.agent.dto;
+package org.apache.inlong.commons.dto;
 
+import java.util.ArrayList;
+import java.util.List;
 import lombok.Data;
+import org.apache.inlong.commons.db.CommandEntity;
 
 @Data
-public class DataConfig {
-    private String inlongGroupId;
-    private String inlongStreamId;
-    private String deliveryTime;
+public class TaskRequestDto {
+    private String agentIp;
     private String uuid;
-    private String ip;
-    private String op;
-    private Integer taskId;
-    private Integer taskType;
-    private String taskConfig;
-
-    public boolean isValid() {
-        return true;
-    }
+    private List<CommandEntity> commandInfo = new ArrayList<>();
 }
